refactor: use multiple implementations to handle writing the PDF in different ways, instead of a single function that works only by storing it on AWS S3
